requirements.txt
openai-agents
pydantic
python-dotenv
eval_type_backport